+++ .mocharc.js
@@ -1,3 +1,4 @@
+'use strict';
 module.exports = {
   exclude: 'test/*/**',
   reporter: 'mocha-multi-reporters'